backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
openai
docling==2.57.0
pdfplumber
docx2txt
docx
pdfkit
jinja2
lxml